#ASLR-28 reorganize project structure
